+++ src/Core/Dfe.Complete.Application/Projects/Models/ConversionTaskDataDto.cs
@@ -27,6 +27,10 @@ namespace Dfe.Complete.Application.Projects.Models
 
         public bool? StakeholderKickOffMeeting { get; set; }
 
+        public bool? StakeholderKickOffCheckProvisionalConversionDate { get; set; }
+
+        public bool? StakeholderKickOffDeclareBudgetChanges { get; set; }
+
         public bool? ConversionGrantCheckVendorAccount { get; set; }
 
         public bool? ConversionGrantPaymentForm { get; set; }
@@ -184,8 +188,6 @@ namespace Dfe.Complete.Application.Projects.Models
         public bool? TenancyAtWillNotApplicable { get; set; }
 
         public bool? TrustModificationOrderNotApplicable { get; set; }
-
-        public bool? StakeholderKickOffCheckProvisionalConversionDate { get; set; }
 
         public bool? ConversionGrantNotApplicable { get; set; }