+++ src/Frontend/Dfe.Complete/Models/ConversionTaskListViewModel.cs
@@ -107,6 +107,7 @@ namespace Dfe.Complete.Models
                  taskData.StakeholderKickOffSetupMeeting == true &&
                  taskData.StakeholderKickOffMeeting == true &&
                  taskData.StakeholderKickOffCheckProvisionalConversionDate == true &&
+                 taskData.StakeholderKickOffDeclareBudgetChanges == true &&
                  project.SignificantDateProvisional == false)
             {
                 return TaskListStatus.Completed;
@@ -115,6 +116,8 @@ namespace Dfe.Complete.Models
                    taskData.StakeholderKickOffLocalAuthorityProforma == true ||
                    taskData.StakeholderKickOffSetupMeeting == true ||
                    taskData.StakeholderKickOffMeeting == true ||
+                   taskData.StakeholderKickOffCheckProvisionalConversionDate == true ||
+                   taskData.StakeholderKickOffDeclareBudgetChanges == true ||
                     project.SignificantDateProvisional == false)
                     ? TaskListStatus.InProgress : TaskListStatus.NotStarted;
         }